Remove FP32 from some layer tests

+++ tests/functional/shared_tests_instances/single_layer_tests/region_yolo.cpp
@@ -25,7 +25,7 @@ namespace {
     };
 
     const std::vector<InferenceEngine::Precision> netPrecisions = {
-            InferenceEngine::Precision::FP32,
+            InferenceEngine::Precision::FP16
     };
 
     INSTANTIATE_TEST_CASE_P(
